LPS-53617 EditEntryAction: move logic to portlet class (no logic changes)
